# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IBLOG AI Assistant",
    "description": "AI Assistant microservice for IBLOG e-learning platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/video/generate": {
      "post": {
        "tags": [
          "video"
        ],
        "summary": "Generate Video",
        "description": "Our side: create a pending job and return immediately -- the caller\npolls GET /jobs/{id} for the result, same pattern as file upload.",
        "operationId": "generate_video_api_v1_video_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/video/jobs": {
      "get": {
        "tags": [
          "video"
        ],
        "summary": "List Jobs",
        "description": "Video worker's side: poll with ?status=pending to claim work.",
        "operationId": "list_jobs_api_v1_video_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VideoJobStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoJobOut"
                  },
                  "title": "Response List Jobs Api V1 Video Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/video/jobs/{job_id}": {
      "get": {
        "tags": [
          "video"
        ],
        "summary": "Get Job",
        "description": "Our side: poll for the result.",
        "operationId": "get_job_api_v1_video_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "video"
        ],
        "summary": "Update Job",
        "description": "Video worker's side: report progress/result. Required fields per\nstatus: 'ready' -> video_url, 'error' -> error_message.",
        "operationId": "update_job_api_v1_video_jobs__job_id__patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoJobUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Language": {
        "type": "string",
        "enum": [
          "fr",
          "en",
          "ar-MA"
        ],
        "title": "Language"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoGenerateRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "description": "Explanation/content to turn into video"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Short topic label for the video's opening frame, in the same language as text"
          },
          "language": {
            "$ref": "#/components/schemas/Language",
            "description": "fr | en | ar-MA"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Chat/quiz session this was triggered from, if any"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id",
            "description": "Company/tenant identifier"
          }
        },
        "type": "object",
        "required": [
          "text",
          "language"
        ],
        "title": "VideoGenerateRequest",
        "description": "Input contract: what our app sends to request a video."
      },
      "VideoJobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "status": {
            "$ref": "#/components/schemas/VideoJobStatus"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "input_text",
          "language",
          "status",
          "created_at"
        ],
        "title": "VideoJobOut"
      },
      "VideoJobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "ready",
          "error"
        ],
        "title": "VideoJobStatus"
      },
      "VideoJobUpdateRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/VideoJobStatus"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url",
            "description": "Required when status='ready'"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Required when status='error'"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "VideoJobUpdateRequest",
        "description": "Output contract: what the video worker PATCHes back as it progresses."
      }
    }
  }
}